+++ app-android/core/model/src/main/java/com/aidatainsight/android/core/model/contract/ContractModels.kt
@@ -8,6 +8,10 @@ object MockApiEnvironment {
     const val DefaultBaseUrl: String = "https://m1.apifoxmock.com/m1/3174267-1700689-default"
 }
 
+object AIChatEndpoint {
+    const val StreamPath: String = "/stream"
+}
+
 @Serializable
 data class ApiEnvironment(
     val name: String,