+++ app-android/core/network/src/test/java/com/aidatainsight/android/core/network/client/AIDataInsightApiClientTest.kt
@@ -125,4 +125,3 @@ private class FakeSessionInvalidationHandler : SessionInvalidationHandler {
         invalidated = true
     }
 }
-